+++ scripts/appveyor/install.ps1
@@ -59,17 +59,12 @@ function InstallPython ($python_version, $architecture, $python_home) {
 
 
 function InstallPip ($python_home) {
-    $pip_path = $python_home + "/Scripts/pip.exe"
     $python_path = $python_home + "/python.exe"
-    if (-not(Test-Path $pip_path)) {
-        Write-Host "Installing pip..."
-        $webclient = New-Object System.Net.WebClient
-        $webclient.DownloadFile($GET_PIP_URL, $GET_PIP_PATH)
-        Write-Host "Executing:" $python_path $GET_PIP_PATH
-        Start-Process -FilePath "$python_path" -ArgumentList "$GET_PIP_PATH" -Wait -Passthru
-    } else {
-        Write-Host "pip already installed."
-    }
+    Write-Host "Installing pip..."
+    $webclient = New-Object System.Net.WebClient
+    $webclient.DownloadFile($GET_PIP_URL, $GET_PIP_PATH)
+    Write-Host "Executing:" $python_path $GET_PIP_PATH
+    Start-Process -FilePath "$python_path" -ArgumentList "$GET_PIP_PATH" -Wait -Passthru
 }
 
 function InstallPackage ($python_home, $pkg) {