+++ HISTORY.rst
@@ -24,6 +24,7 @@ PlatformIO 2.0
   `atmelavr <http://docs.platformio.org/en/latest/platforms/atmelavr.html#boards>`__
   platform
   (`pull #256 <https://github.com/platformio/platformio/pull/256>`_)
+* Added short ``-h`` help option for PlatformIO and sub-commands
 * Updated ``tool-teensy`` for `Teensy <http://docs.platformio.org/en/latest/platforms/teensy.html>`__
   platform
   (`issue #268 <https://github.com/platformio/platformio/issues/268>`_)